refactor(poster): Move scrollanimation to the base
